refactor(workflow): consolidate BaseNodeData and related entities
- Moved BaseNodeData and related classes from `base_node_data_entities` to `nodes.base.entities`.
- Updated imports across various files to reflect new location of BaseNodeData.
- Created `event` module for handling NodeEvents to improve organization.
- Optimized usage of TYPE_CHECKING for conditional imports, enhancing performance.
- Improved class metadata management by assigning default factories and utilizing type annotations.
