+++ api/core/app/entities/queue_entities.py
@@ -5,9 +5,9 @@ from typing import Any, Optional
 from pydantic import BaseModel, field_validator
 
 from core.model_runtime.entities.llm_entities import LLMResult, LLMResultChunk
-from core.workflow.entities.base_node_data_entities import BaseNodeData
 from core.workflow.entities.node_entities import NodeRunMetadataKey
 from core.workflow.graph_engine.entities.graph_runtime_state import GraphRuntimeState
+from core.workflow.nodes.base import BaseNodeData
 from enums import NodeType
