+++ api/core/workflow/entities/base_node_data_entities.py
@@ -1,16 +1,4 @@
-from abc import ABC
-from typing import Optional
-
 from pydantic import BaseModel
-
-
-class BaseNodeData(ABC, BaseModel):
-    title: str
-    desc: Optional[str] = None
-
-
-class BaseIterationNodeData(BaseNodeData):
-    start_node_id: Optional[str] = None
 
 
 class BaseIterationState(BaseModel):
