+++ api/core/workflow/graph_engine/entities/event.py
@@ -3,8 +3,8 @@ from typing import Any, Optional
 
 from pydantic import BaseModel, Field
 
-from core.workflow.entities.base_node_data_entities import BaseNodeData
 from core.workflow.graph_engine.entities.runtime_route_state import RouteNodeState
+from core.workflow.nodes.base import BaseNodeData
 from enums import NodeType
