+++ api/core/workflow/nodes/base/__init__.py
@@ -1,3 +1,4 @@
+from .entities import BaseIterationNodeData, BaseIterationState, BaseNodeData
 from .node import BaseNode
 
-__all__ = ["BaseNode"]
+__all__ = ["BaseNode", "BaseNodeData", "BaseIterationNodeData", "BaseIterationState"]
