+++ api/core/workflow/nodes/end/entities.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel, Field
 
-from core.workflow.entities.base_node_data_entities import BaseNodeData
 from core.workflow.entities.variable_entities import VariableSelector
+from core.workflow.nodes.base import BaseNodeData
 
 
 class EndNodeData(BaseNodeData):
