+++ api/core/workflow/nodes/event/types.py
@@ -1,0 +1,3 @@
+from .event import RunCompletedEvent, RunRetrieverResourceEvent, RunStreamChunkEvent
+
+NodeEvent = RunCompletedEvent | RunStreamChunkEvent | RunRetrieverResourceEvent
