+++ api/core/workflow/nodes/if_else/entities.py
@@ -2,7 +2,7 @@ from typing import Literal, Optional
 
 from pydantic import BaseModel, Field
 
-from core.workflow.entities.base_node_data_entities import BaseNodeData
+from core.workflow.nodes.base import BaseNodeData
 from core.workflow.utils.condition.entities import Condition
