+++ api/core/workflow/nodes/iteration/entities.py
@@ -1,6 +1,8 @@
 from typing import Any, Optional
 
-from core.workflow.entities.base_node_data_entities import BaseIterationNodeData, BaseIterationState, BaseNodeData
+from pydantic import Field
+
+from core.workflow.nodes.base import BaseIterationNodeData, BaseIterationState, BaseNodeData
 
 
 class IterationNodeData(BaseIterationNodeData):
@@ -26,7 +28,7 @@ class IterationState(BaseIterationState):
     Iteration State.
     """
 
-    outputs: list[Any] = None
+    outputs: list[Any] = Field(default_factory=list)
     current_output: Optional[Any] = None
 
     class MetaData(BaseIterationState.MetaData):
