+++ api/core/workflow/nodes/knowledge_retrieval/entities.py
@@ -2,7 +2,7 @@ from typing import Any, Literal, Optional
 
 from pydantic import BaseModel
 
-from core.workflow.entities.base_node_data_entities import BaseNodeData
+from core.workflow.nodes.base import BaseNodeData
 
 
 class RerankingModelConfig(BaseModel):
