+++ api/core/workflow/nodes/list_operator/entities.py
@@ -3,7 +3,7 @@ from typing import Literal
 
 from pydantic import BaseModel, Field
 
-from core.workflow.entities.base_node_data_entities import BaseNodeData
+from core.workflow.nodes.base import BaseNodeData
 
 _Condition = Literal[
     # string conditions
