+++ api/core/workflow/nodes/loop/entities.py
@@ -1,4 +1,4 @@
-from core.workflow.entities.base_node_data_entities import BaseIterationNodeData, BaseIterationState
+from core.workflow.nodes.base import BaseIterationNodeData, BaseIterationState
 
 
 class LoopNodeData(BaseIterationNodeData):
