+++ api/core/workflow/nodes/start/entities.py
@@ -3,7 +3,7 @@ from collections.abc import Sequence
 from pydantic import Field
 
 from core.app.app_config.entities import VariableEntity
-from core.workflow.entities.base_node_data_entities import BaseNodeData
+from core.workflow.nodes.base import BaseNodeData
 
 
 class StartNodeData(BaseNodeData):
