+++ api/core/workflow/nodes/template_transform/entities.py
@@ -1,5 +1,5 @@
-from core.workflow.entities.base_node_data_entities import BaseNodeData
 from core.workflow.entities.variable_entities import VariableSelector
+from core.workflow.nodes.base import BaseNodeData
 
 
 class TemplateTransformNodeData(BaseNodeData):
