+++ api/core/workflow/nodes/variable_assigner/node.py
@@ -2,9 +2,8 @@ from sqlalchemy import select
 from sqlalchemy.orm import Session
 
 from core.variables import SegmentType, Variable
-from core.workflow.entities.base_node_data_entities import BaseNodeData
 from core.workflow.entities.node_entities import NodeRunResult
-from core.workflow.nodes.base import BaseNode
+from core.workflow.nodes.base import BaseNode, BaseNodeData
 from enums import NodeType
 from extensions.ext_database import db
 from factories import variable_factory
